requirements.txt
beautifulsoup4==4.12.2
cffi==1.16.0
cryptography==41.0.5
IMAPClient==2.3.1
pycparser==2.21
PyMySQL==1.1.0
PyYAML==6.0.1
pyzmail36==1.0.5
schedule==1.2.1
six==1.16.0
soupsieve==2.5
